# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Compte API",
        "version": "v1",
        "description": "Documentation minimale — ajoutez ici vos endpoints OpenAPI."
    },
    "servers": [
        { "url": "http://localhost:8000", "description": "Local development (use /api/v1 paths)" },
        { "url": "https://compte-api-hjvy.onrender.com/khouss.ngom", "description": "Hosted docs (Render) — production prefix /khouss.ngom (khouss.ngom)" },
        { "url": "/", "description": "Use same origin as Swagger UI (recommended for hosted docs)" }
    ],
    "paths": {
        "/api/v1/comptes/{id}/archive": {
            "post": {
                "summary": "Archiver un compte (marquer comme archived)",
                "tags": ["Comptes"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "ID numérique du compte à archiver" }
                ],
                "responses": {
                    "200": {
                        "description": "Compte archivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/ArchiveResponse" }
                            }
                        }
                    },
                    "403": { "description": "Accès refusé" },
                    "404": { "description": "Compte introuvable" }
                }
            }
        },
            "/api/v1/comptes": {
            "get": {
                "summary": "Lister tous les comptes",
                "description": "Récupère une liste paginée de comptes avec possibilité de filtrage et tri",
                "tags": ["Comptes"],
                "parameters": [
                    { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "Numéro de page" },
                    { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Nombre d'éléments par page" },
                    { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["epargne", "cheque"] }, "description": "Filtrer par type de compte" },
                    { "name": "statut", "in": "query", "schema": { "type": "string", "enum": ["actif", "bloque", "ferme"] }, "description": "Filtrer par statut" },
                    { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Recherche par numéro de compte ou nom du titulaire" },
                    { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["date_creation", "solde", "titulaire_compte"], "default": "date_creation" }, "description": "Champ de tri (utilisez les clés exactes: date_creation, solde, titulaire_compte)" },
                    { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }, "description": "Ordre de tri" }
                ],
                "responses": {
                "200": {
                    "description": "Liste des comptes récupérée avec succès",
                    "content": {
                    "application/json": {
                        "schema": { "$ref": "#/components/schemas/ComptesResponse" }
                    }
                    }
                }
                }
            }
            },
...
            "/api/v1/comptes/{numeroCompte}": {
                "get": {
                    "summary": "Récupérer un compte par son numéro (Admin ou Client)",
                    "tags": ["Comptes"],
                    "parameters": [
                        { "name": "numeroCompte", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Numéro de compte (numero_compte)" }
                    ],
                    "responses": {
                        "200": {
                            "description": "Compte récupéré",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": { "type": "boolean" },
                                            "data": { "$ref": "#/components/schemas/Compte" }
                                        }
                                    }
                                }
                            }
                        },
                        "403": { "description": "Accès refusé" },
                        "404": {
                            "description": "Compte introuvable",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "type": "object",
                                        "properties": {
                                            "success": { "type": "boolean", "example": false },
                                            "error": {
                                                "type": "object",
                                                "properties": {
                                                    "code": { "type": "string", "example": "COMPTE_NOT_FOUND" },
                                                    "message": { "type": "string", "example": "Le compte C00123456 est introuvable" }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
...
        },
        "components": {
            "schemas": {
            "Compte": {
                "type": "object",
                "properties": {
                "id": { "type": "string", "format": "uuid" },
                "numeroCompte": { "type": "string" },
                "titulaire": { "type": "string" },
                "type": { "type": "string", "enum": ["epargne", "cheque"] },
                "solde": { "type": "number" },
                "devise": { "type": "string" },
                "dateCreation": { "type": "string", "format": "date-time" },
                "statut": { "type": "string", "enum": ["actif", "bloque", "ferme"] },
                "dateDebutBlocage": { "type": "string", "format": "date", "nullable": true },
                "dateFinBlocage": { "type": "string", "format": "date", "nullable": true },
                "motifBlocage": { "type": "string", "nullable": true },
                "archived": { "type": "boolean", "nullable": true },
                "metadonnees": { "type": "object", "properties": { "derniereModification": { "type": "string", "format": "date-time" }, "version": { "type": "integer" } } }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                "currentPage": { "type": "integer" },
                "totalPages": { "type": "integer" },
                "totalItems": { "type": "integer" },
                "itemsPerPage": { "type": "integer" },
                "hasNext": { "type": "boolean" },
                "hasPrevious": { "type": "boolean" }
                }
            },
            "Links": {
                "type": "object",
                "properties": {
                "self": { "type": "string" },
                "next": { "type": "string" },
                "first": { "type": "string" },
                "last": { "type": "string" }
                }
            },
            "ComptesResponse": {
                "type": "object",
                "properties": {
                "success": { "type": "boolean" },
                "data": { "type": "array", "items": { "$ref": "#/components/schemas/Compte" } },
                "pagination": { "$ref": "#/components/schemas/Pagination" },
                "links": { "$ref": "#/components/schemas/Links" }
                }
            },
...
            "ArchiveResponse": {
                "type": "object",
                "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Compte archivé avec succès" },
                    "data": { "$ref": "#/components/schemas/Compte" }
                }
            }
            }
        }
}